Rough Combat System implemented and others!
Have a rough system made for combat where Fire1, or left mouse click by default, fires a shuriken that when it hits an enemy deals 10 damage. If the enemy takes damage that puts them below 1HP, they get destroyed. Also started making prefabs for the scenery with Bowen's sprite work.

+++ Assets/Scripts/LifeTime.cs
@@ -6,7 +6,7 @@ public class LifeTime : MonoBehaviour
 {
     // Start is called before the first frame update
 
-    public int life = 2;
+    public float life = 2;
     void Start()
     {
         Destroy(gameObject, life);
